+++ packages/riverpod_cli/CHANGELOG.md
@@ -1,5 +1,7 @@
-# 0.0.2+2
+# 0.0.3
+- Migration for unifying syntax (riverpod 1.0.0) [RFC](https://github.com/rrousselGit/river_pod/issues/335)
 
+# 0.0.2+2
 - Fix migration for static final providers.
 
 # 0.0.2+1